requirements.txt
aiofiles==0.8.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
asyncio==3.4.3
asyncpraw==7.7.1
asyncprawcore==2.4.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
discord.py==2.4.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-search-results==2.4.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.22.0
huggingface_hub==0.26.2
idna==3.10
multidict==6.1.0
packaging==24.2
pillow==11.0.0
praw==7.7.1
prawcore==2.4.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.1
pydantic_core==2.27.1
pyparsing==3.2.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
serpapi==0.1.5
tqdm==4.67.0
typing==3.7.4.3
typing_extensions==4.12.2
Unidecode==1.3.8
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
websocket-client==1.8.0
Wikipedia-API==0.6.0
yarl==1.18.0
